// front-end/src/app/Telas/usuario/models/vagasUsuario.ts
export class VagasUsuario {
    id: string;
    adminId: string;
    funcao: string;
    descricao: string;
    area: string;
    horario: string;
    salario: number;
    beneficios: string;
    qtdVagas: number;
    empresa: string;    
}

export class DadosUsuario {
    id: string;
    nome: string;
    email: string;
    dataNascimento: string;
    profissao: string;
    experiencia: string;
    cursos: string;    
}